requirements.txt
flask>=2.3.0
Pillow>=10.0.0
numpy>=2.0.0
pytesseract==0.3.8
music21==7.1.0
opencv-python>=4.8.0